@@ -1,7 +1,7 @@
 {
     "name": "@harbor/ui",
     "version": "1.9.0",
-    "description": "Harbor shared UI components based on Clarity and Angular7",
+    "description": "Harbor shared UI components based on Clarity and Angular8",
     "author": "CNCF",
     "module": "index.js",
     "main": "bundles/harborui.umd.min.js",
@@ -10,7 +10,7 @@
     "keywords": [
         "Harbor",
         "Clarity",
-        "Angular7"
+        "Angular8"
     ],
     "license": "Apache 2.0",
     "repository": {
@@ -19,30 +19,29 @@
     },
     "homepage": "https://github.com/vmware/harbor#readme",
     "peerDependencies": {
-        "@angular/animations": "^7.1.3",
-        "@angular/common": "^7.1.3",
-        "@angular/compiler": "^7.1.3",
-        "@angular/core": "^7.1.3",
-        "@angular/forms": "^7.1.3",
-        "@angular/http": "^7.1.3",
-        "@angular/platform-browser": "^7.1.3",
-        "@angular/platform-browser-dynamic": "^7.1.3",
-        "@angular/router": "^7.1.3",
+        "@angular/animations": "~8.2.0",
+        "@angular/common": "~8.2.0",
+        "@angular/compiler": "~8.2.0",
+        "@angular/core": "~8.2.0",
+        "@angular/forms": "~8.2.0",
+        "@angular/platform-browser": "~8.2.0",
+        "@angular/platform-browser-dynamic": "~8.2.0",
+        "@angular/router": "~8.2.0",
         "@ngx-translate/core": "^10.0.2",
         "@ngx-translate/http-loader": "^3.0.1",
-        "@webcomponents/custom-elements": "^1.1.3",
-        "@clr/angular": "^1.0.0",
-        "@clr/ui": "^1.0.0",
-        "@clr/icons": "^1.0.0",
+        "@webcomponents/custom-elements": "^1.2.4",
+        "@clr/angular": "^2.1.0",
+        "@clr/ui": "^2.1.0",
+        "@clr/icons": "^2.1.0",
         "core-js": "^2.5.4",
         "intl": "^1.2.5",
         "mutationobserver-shim": "^0.3.2",
         "ngx-cookie": "^1.0.0",
-        "ngx-markdown": "^6.2.0",
-        "rxjs": "^6.3.3",
+        "ngx-markdown": "^7.1.5",
         "ts-helpers": "^1.1.1",
-        "web-animations-js": "^2.2.1",
-        "zone.js": "^0.8.26",
-        "tslib": "^1.9.0"
+        "web-animations-js": "^2.3.2",
+        "rxjs": "~6.5.2",
+        "tslib": "^1.9.0",
+        "zone.js": "~0.9.1"
     }
 }
