@@ -34,7 +34,7 @@ export class VulnerabilityConfigComponent implements OnInit {
     openState: boolean = false;
     getLabelCurrent: string;
 
-    @ViewChild(CronScheduleComponent)
+    @ViewChild(CronScheduleComponent, { static: false })
     CronScheduleComponent: CronScheduleComponent;
 
     @Input()
@@ -109,7 +109,7 @@ export class VulnerabilityConfigComponent implements OnInit {
           };
         }
       }
-    @ViewChild("systemConfigFrom") systemSettingsForm: NgForm;
+    @ViewChild("systemConfigFrom", { static: false }) systemSettingsForm: NgForm;
 
     get isValid(): boolean {
         return this.systemSettingsForm && this.systemSettingsForm.valid;
