@@ -46,8 +46,8 @@ export class AddRobotComponent implements OnInit, OnDestroy {
   @Input() projectId: number;
   @Input() projectName: string;
   @Output() create = new EventEmitter<boolean>();
-  @ViewChild("robotForm") currentForm: NgForm;
-  @ViewChild("copyAlert") copyAlert: InlineAlertComponent;
+  @ViewChild("robotForm", { static: false }) currentForm: NgForm;
+  @ViewChild("copyAlert", { static: false }) copyAlert: InlineAlertComponent;
   constructor(
       private robotService: RobotService,
       private translate: TranslateService,
